refactor: improve database service organization
- Split DatabaseService into smaller services:
  - ProjectDatabaseService: project CRUD operations
  - ChatDatabaseService: chat message operations
  - SaveDatabaseService: project state save/load operations
- Created main DatabaseService to combine all services
- Improved code organization and maintainability
