+++ apps/backend/src/services/database/index.ts
@@ -1,0 +1,4 @@
+export { dbService } from './DatabaseService';
+export { ProjectDatabaseService } from './ProjectDatabaseService';
+export { ChatDatabaseService } from './ChatDatabaseService';
+export { SaveDatabaseService } from './SaveDatabaseService';
